requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-cryptography==1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-decouple==3.5
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
zope.interface==5.4.0
